.ipynb_checkpoints/requirements-checkpoint.txt
Flask==2.3.3
tensorflow==2.13.0
opencv-python==4.8.0.74
Pillow==10.0.0
numpy==1.24.3
Werkzeug==2.3.7
